[DC][REL BUILD] Disable Debugging, Chat is default
This bug disable debugging when the build type is release. With this
user won't be able to access developer options. Even rooted device won't
allow use to gain access to the APP.

This commit also makes DC allow Chat only option. Not user won't get
email notification on the APP.

Bug: #21 & #25
Test: Verified on Emulator and physical phone.