[DC] Delete msg from chatlist & chat
Now, user can delete the chat from UI chatlist as well as from chat
itself. This doesnot delete the chat from the peer and this is as per
the design. However if the add request come for the same user, then the
peer should flush all the data and reinitiate the communication again
starting from KGM handshake.

Will open a seperate issue #51

Bug: #33
Test: Verified on Emulator and physical phone.